+++ crates/kitsune-core/Cargo.toml
@@ -6,10 +6,6 @@ version.workspace = true
 [dependencies]
 ammonia = "3.3.0"
 argon2 = "0.5.2"
-askama = { version = "0.12.0", features = [
-    "with-axum",
-], default-features = false }
-askama_axum = "0.3.0"
 async-recursion = "1.0.5"
 async-stream = "0.3.5"
 async-trait = "0.1.73"
@@ -54,9 +50,6 @@ kitsune-storage = { path = "../kitsune-storage" }
 kitsune-type = { path = "../kitsune-type" }
 mime = "0.3.17"
 mime_guess = { version = "2.0.4", default-features = false }
-oxide-auth = "0.5.3"
-oxide-auth-async = "0.1.0"
-oxide-auth-axum = "0.3.0"
 password-hash = { version = "0.5.0", features = ["std"] }
 pin-project-lite = "0.2.13"
 pkcs8 = { version = "0.10.2", features = ["std"] }
