+++ crates/kitsune-core/src/state.rs
@@ -2,19 +2,10 @@ use crate::{
     activitypub::Fetcher,
     event::PostEventEmitter,
     service::{
-        account::AccountService,
-        attachment::AttachmentService,
-        captcha::CaptchaService,
-        federation_filter::FederationFilterService,
-        instance::InstanceService,
-        job::JobService,
-        mailing::MailingService,
-        notification::NotificationService,
-        oauth2::{OAuth2Service, OAuthEndpoint},
-        post::PostService,
-        timeline::TimelineService,
-        url::UrlService,
-        user::UserService,
+        account::AccountService, attachment::AttachmentService, captcha::CaptchaService,
+        federation_filter::FederationFilterService, instance::InstanceService, job::JobService,
+        mailing::MailingService, notification::NotificationService, post::PostService,
+        timeline::TimelineService, url::UrlService, user::UserService,
     },
     webfinger::Webfinger,
 };
@@ -48,7 +39,6 @@ impl_from_ref! {
         FederationFilterService => |input: &Zustand| input.service.federation_filter.clone(),
         JobService => |input: &Zustand| input.service.job.clone(),
         NotificationService => |input: &Zustand| input.service.notification.clone(),
-        OAuth2Service => |input: &Zustand| input.service.oauth2.clone(),
         PostService => |input: &Zustand| input.service.post.clone(),
         SearchService => |input: &Zustand| input.service.search.clone(),
         InstanceService => |input: &Zustand| input.service.instance.clone(),
@@ -129,7 +119,6 @@ pub struct Service {
     pub job: JobService,
     pub mailing: MailingService,
     pub notification: NotificationService,
-    pub oauth2: OAuth2Service,
     #[cfg(feature = "oidc")]
     pub oidc: Option<OidcService>,
     pub post: PostService,
@@ -152,7 +141,6 @@ pub struct Zustand {
     pub fetcher: Fetcher,
     #[cfg(feature = "mastodon-api")]
     pub mastodon_mapper: crate::mapping::MastodonMapper,
-    pub oauth_endpoint: OAuthEndpoint,
     pub service: Service,
     pub session_config: SessionConfig,
     pub webfinger: Webfinger,
