+++ kitsune/src/http/handler/mastodon/api/v1/accounts/relationships.rs
@@ -20,7 +20,7 @@ pub struct RelationshipQuery {
     id: Vec<Uuid>,
 }
 
-#[debug_handler(state = crate::state::Zustand)]
+#[debug_handler(state = kitsune_core::state::Zustand)]
 #[utoipa::path(
     get,
     path = "/api/v1/accounts/relationships",
