+++ kitsune/src/http/handler/mastodon/api/v1/accounts/unfollow.rs
@@ -12,7 +12,7 @@ use axum::{
 use kitsune_type::mastodon::relationship::Relationship;
 use speedy_uuid::Uuid;
 
-#[debug_handler(state = crate::state::Zustand)]
+#[debug_handler(state = kitsune_core::state::Zustand)]
 pub async fn post(
     State(account_service): State<AccountService>,
     State(mastodon_mapper): State<MastodonMapper>,
