+++ kitsune/src/http/handler/mastodon/api/v1/follow_requests/reject.rs
@@ -13,7 +13,7 @@ use crate::{
     service::account::{AccountService, FollowRequest},
 };
 
-#[debug_handler(state = crate::state::Zustand)]
+#[debug_handler(state = kitsune_core::state::Zustand)]
 #[utoipa::path(
     post,
     path = "/api/v1/follow_requests/{id}/reject",
