+++ kitsune/src/http/handler/mastodon/api/v1/statuses/favourite.rs
@@ -12,7 +12,7 @@ use axum::{
 use kitsune_type::mastodon::Status;
 use speedy_uuid::Uuid;
 
-#[debug_handler(state = crate::state::Zustand)]
+#[debug_handler(state = kitsune_core::state::Zustand)]
 #[utoipa::path(
     post,
     path = "/api/v1/statuses/{id}/favourite",
