+++ kitsune/src/http/handler/mastodon/api/v1/statuses/reblog.rs
@@ -20,7 +20,7 @@ pub struct RepostBody {
     visibility: Visibility,
 }
 
-#[debug_handler(state = crate::state::Zustand)]
+#[debug_handler(state = kitsune_core::state::Zustand)]
 #[utoipa::path(
     post,
     path = "/api/v1/statuses/{id}/reblog",
