+++ kitsune/src/http/handler/oauth/mod.rs
@@ -2,11 +2,11 @@
 //! Standard-compliant OAuth2 flows
 //!
 
-use crate::state::Zustand;
 use axum::{
     routing::{get, post},
     Router,
 };
+use kitsune_core::state::Zustand;
 
 pub mod authorize;
 pub mod token;
