+++ kitsune/src/http/handler/posts/activity.rs
@@ -1,10 +1,9 @@
-use crate::{
-    error::Result, http::responder::ActivityPubJson, mapping::IntoActivity, state::Zustand,
-};
+use crate::{error::Result, http::responder::ActivityPubJson};
 use axum::{
     debug_handler,
     extract::{Path, State},
 };
+use kitsune_core::{mapping::IntoActivity, state::Zustand};
 use kitsune_type::ap::Activity;
 use speedy_uuid::Uuid;
