+++ kitsune/src/http/handler/users/following.rs
@@ -1,9 +1,8 @@
-use crate::{
-    error::Result, http::responder::ActivityPubJson, service::url::UrlService, state::Zustand,
-};
+use crate::{error::Result, http::responder::ActivityPubJson};
 use axum::extract::{OriginalUri, Path, State};
 use diesel::{BoolExpressionMethods, ExpressionMethods, JoinOnDsl, QueryDsl};
 use diesel_async::RunQueryDsl;
+use kitsune_core::{service::url::UrlService, state::Zustand};
 use kitsune_db::schema::{accounts, accounts_follows};
 use kitsune_type::ap::{
     ap_context,
