+++ kitsune/src/http/handler/users/outbox.rs
@@ -1,15 +1,14 @@
-use crate::{
-    error::Result,
-    http::responder::ActivityPubJson,
-    mapping::IntoActivity,
-    service::{account::GetPosts, url::UrlService},
-    state::Zustand,
-};
+use crate::{error::Result, http::responder::ActivityPubJson};
 use axum::extract::{OriginalUri, Path, Query, State};
 use axum_extra::either::Either;
 use diesel::{BelongingToDsl, ExpressionMethods, QueryDsl, SelectableHelper};
 use diesel_async::RunQueryDsl;
 use futures_util::{stream, StreamExt, TryStreamExt};
+use kitsune_core::{
+    mapping::IntoActivity,
+    service::{account::GetPosts, url::UrlService},
+    state::Zustand,
+};
 use kitsune_db::{
     model::{account::Account, post::Post},
     post_permission_check::{PermissionCheck, PostPermissionCheckExt},
