+++ kitsune/src/http/handler/well_known/nodeinfo.rs
@@ -1,5 +1,5 @@
-use crate::{service::url::UrlService, state::Zustand};
 use axum::{extract::State, routing, Json, Router};
+use kitsune_core::{service::url::UrlService, state::Zustand};
 use kitsune_type::nodeinfo::well_known::{Link, WellKnown};
 
 #[allow(clippy::unused_async)]
