+++ kitsune/src/oauth2/registrar.rs
